+++ trell-android/app/src/main/java/com/example/trellandroid/data/responses/UserResponse.kt
@@ -1,7 +1,7 @@
 package com.example.trellandroid.data.responses
 
 data class UserResponse(
-    val userId: Double? = null,
+    val userId: Long? = null,
     val username: String? = null,
     val name: String? = null,
     val totalPosts: Int = 0,
